requirements.txt
adbutils==0.4.4
Appium-Python-Client==0.46
attrs==19.1.0
baidu-aip==2.2.18.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
deprecation==2.0.7
facebook-wda==0.3.9
humanize==0.5.1
idna==2.8
logzero==1.5.0
lxml==4.4.1
numpy==1.18.4
opencv-python==4.2.0.34
packaging==19.1
Pillow==6.1.0
progress==1.5
py==1.8.0
pyparsing==2.4.2
requests==2.22.0
retry==0.9.2
selenium==3.141.0
six==1.12.0
tornado==6.0.3
uiautomator2==1.2.4.dev2
urllib3==1.25.3
weditor==0.2.5
whichcraft==0.6.0
